+++ google/cloud/developerconnect/quickstart/quickstart.cc
@@ -18,12 +18,12 @@
 #include <iostream>
 
 int main(int argc, char* argv[]) try {
-  if (argc != 3) {
-    std::cerr << "Usage: " << argv[0] << " project-id location-id\n";
+  if (argc != 2) {
+    std::cerr << "Usage: " << argv[0] << " project-id\n";
     return 1;
   }
 
-  auto const location = google::cloud::Location(argv[1], argv[2]);
+  auto const location = google::cloud::Location(argv[1], "global");
 
   namespace developerconnect = ::google::cloud::developerconnect_v1;
   auto client = developerconnect::DeveloperConnectClient(